[fix] api 테스트 코드 작성

+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqModifyMessage.java
@@ -4,6 +4,6 @@ package com.tosslab.jandi.app.network.models;
  * Created by justinygchoi on 2014. 5. 28..
  */
 public class ReqModifyMessage {
-    public int teamId;
+    public long teamId;
     public String content;
 }
